Refactor poll validators into utils
This will let us use the same validator createPoll as well as on the lfrontend if we want to in the future
